updated term 'dosage' to 'dose' for correct usage. Renamed 'cumulative_drug_dose' to 'actual_drug_dose'.
