+++ schemas/chemotherapy.json
@@ -78,52 +78,50 @@
       }
     },
     {
-      "name": "chemotherapy_dosage_units",
-      "description": "Indicate units used to record chemotherapy drug dosage.",
+      "name": "chemotherapy_drug_dose_units",
+      "description": "Indicate units used to record chemotherapy drug dose.",
       "valueType": "string",
       "restrictions": {
         "required": true,
-        "codeList": "#/list/drug_dosage_units"
+        "codeList": "#/list/drug_dose_units"
       },
       "meta": {
         "core": true,
-        "displayName": "Chemotherapy Dosage Units"
+        "displayName": "Chemotherapy Drug Dose Units"
       }
     },
     {
-      "name": "prescribed_cumulative_drug_dosage",
-      "description": "Indicate the total prescribed drug dose in the same units specified in chemotherapy_dosage_units.",
+      "name": "prescribed_cumulative_drug_dose",
+      "description": "Indicate the total prescribed drug dose in the same units specified in chemotherapy_drug_dose_units.",
       "valueType": "number",
       "restrictions": {
-        "required": true,
-        "script": "#/script/chemotherapy/drugDosage",
+        "script": "#/script/chemotherapy/drugDose",
         "range": {
            "exclusiveMin": 0
         }
       },
       "meta": {
         "core": true,
-        "displayName": "Prescribed Cumulative Drug Dosage",
-        "dependsOn": "chemotherapy.cumulative_drug_dosage",
-        "notes": "Either the actual drug dosage (cumulative_drug_dosage) or the prescribed drug dosage (prescribed_cumulative_drug_dosage) field must be submitted."
+        "displayName": "Prescribed Cumulative Drug Dose",
+        "dependsOn": "chemotherapy.actual_cumulative_drug_dose",
+        "notes": "Either the 'actual_cumulative_drug_dose' or the 'prescribed_cumulative_drug_dose' field must be submitted."
       }
     },
     {
-      "name": "cumulative_drug_dosage",
-      "description": "Indicate the total actual drug dose in the same units specified in chemotherapy_dosage_units.",
+      "name": "actual_cumulative_drug_dose",
+      "description": "Indicate the total actual drug dose in the same units specified in chemotherapy_drug_dose_units.",
       "valueType": "number",
       "restrictions": {
-        "required": true,
-        "script": "#/script/chemotherapy/drugDosage",
+        "script": "#/script/chemotherapy/drugDose",
         "range": {
            "exclusiveMin": 0
         }
       },
       "meta": {
         "core": true,
-        "displayName": "Actual Cumulative Drug Dosage",
-        "dependsOn": "chemotherapy.prescribed_cumulative_drug_dosage",
-        "notes": "Either the actual drug dosage (cumulative_drug_dosage) or the prescribed drug dosage (prescribed_cumulative_drug_dosage) field must be submitted."
+        "displayName": "Actual Cumulative Drug Dose",
+        "dependsOn": "chemotherapy.prescribed_cumulative_drug_dose",
+        "notes": "Either the 'actual_cumulative_drug_dose' or the 'prescribed_cumulative_drug_dose' field must be submitted."
       }
     },
     {
